backend/requirements.txt
fastapi==0.103.2
uvicorn==0.23.2
pyodbc
pandas
python-dotenv==1.0.0
